backend/requirements.txt
fastapi==0.109.1
uvicorn==0.24.0

langchain==1.2.0
langchain-community==0.4.1
langchain-classic==1.0.0
langchain-core==1.2.3
langchain-text-splitters==1.1.0

sentence-transformers==2.2.2
torch==2.1.0
transformers==4.33.0
huggingface-hub==0.16.4

faiss-cpu==1.7.4
numpy==1.26.4

pypdf>=6.4.0
python-multipart==0.0.6

python-dotenv==1.0.0